backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
pillow==11.1.0
ollama==0.4.*